% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countem.R
\name{countem}
\alias{countem}
\title{Cycle through bar charts of your data frame}
\usage{
countem(
  data,
  chart_interval = 3500,
  bar_interval = 2500,
  sort_bars = TRUE,
  include_ordinal = TRUE,
  fill = "steelblue"
)
}
\arguments{
\item{data}{A table of data}

\item{chart_interval}{How quickly to switch charts, in milliseconds}

\item{bar_interval}{How quickly the bars transition, in milliseconds}

\item{sort_bars}{Whether to sort the bars of the bar chart}

\item{include_ordinal}{Whether to include ordinal variables}

\item{fill}{Fill color of bars}
}
\description{
Cycle through bar charts of your data frame
}
\examples{
\dontrun{
countem(ggplot2::diamonds)
}
}
